+++ lib2/index.ts
@@ -1,3 +1,0 @@
-export { createDrill } from '../lib/drill'
-export { createChain } from './chains'
-export { brand, custom, scope } from './utils'